fix: workaround integer type loss from crossplane using structpb.Struct for unstructured.Unstructured composite
structpb.Struct is the obvious choide for dynamic typed struct fields in protobuf.
Unfortunately it causes all numbers to be cast to float64 which serde won't accept for i64 fields like metadata.generation.

As a workaround we adjust the intermediate JSON representartion when translating from the crossplane Resource
to our custom type and patch every float that can be interpreted as i64 without data loss as i64.