feat!: rework k8s-openapi dependency handling in sdk
Following the `k8s-openapi` upstream recommendation to not set any `v1_*`
features for the `k8s-openapi`-crate in libraries that utilize that
but to leave this to binary applications to decide.

Therefore the `k8s-openapi`-crate is no longer re-exported by the sdk.
Furthermore, the sdk-implementation of the `TryFromResource`- and `TryIntoResource`-create
for types that implement `k8s_openapi::Metadata` have been associated with a `k8s-openapi`-feature
to allow opting-out of the `k8s-openapi`-sdk-dependency.