feat!: make `Args` public, remove `tracing` re-export from sdk
The sdk never really did much in terms of setting up tracing (basically just
setting the log level and using json, which should be left to the user).

By exporting arg and expecting the user to provide it back to `launch_server`
the sdk user can use it e.g. to setup logging with the logging framework they prefer.
Furthermore, for users that don't want to use `Args` there is basically just an additional
`Args::parse()` argument that has to be added to launch-server`.

- Tracing library is no longer re-exporter
- Signature of `launch_server` has changed. The original `impl CompositeFunction` argument is now
  the second instead of the only argument. The first argument are the cli arguments which can be obtained
  via `Args::parse()`.